using byte on shell instead int

+++ src/main/org/x96/sys/foundation/cs/lexer/visitors/crafted/book/terminals/docOrSpace/DocOrEmptySpace.java
@@ -4,6 +4,7 @@ import org.x96.sys.foundation.cs.ast.book.passage.pattern.modifier.Shell;
 import org.x96.sys.foundation.cs.lexer.router.switcher.Switcher;
 import org.x96.sys.foundation.cs.lexer.token.Token;
 import org.x96.sys.foundation.cs.lexer.tokenizer.Tokenizer;
+
 import org.x96.sys.foundation.cs.lexer.visitors.crafted.book.terminals.doc.Doc;
 import org.x96.sys.foundation.cs.lexer.visitors.crafted.book.terminals.emptySpace.EmptySpace;
 
@@ -23,7 +24,7 @@ public class DocOrEmptySpace extends Doc {
         switcher.know(EmptySpace.class);
         switcher.know(Doc.class);
         push(switcher.stream(tokenizer));
-        setMod(new Shell(0x40));
+        setMod(new Shell((byte) 0x40));
         return stream();
     }
